--- extension/ss_client/chinadns
@@ -1,1 +1,0 @@
-bin.armv7l/chinadns
