--- extension/ss_client/fw_ss_client
@@ -1,1 +1,0 @@
-bin.armv7l/fw_ss_client
