--- extension/ss_client/fw_ss_redir
@@ -1,1 +1,0 @@
-bin.armv7l/fw_ss_redir
