--- extension/ss_client/ss_client.service
@@ -1,13 +1,0 @@
-[Unit]
-Description=Firewalla SS Client
-After=network.target
-
-[Service]
-User=root
-Type=simple
-Restart=always
-WorkingDirectory=/home/pi/firewalla/
-ExecStart=/bin/bash /home/pi/firewalla/extension/ss_client/ss_client.sh
-
-[Install]
-WantedBy=multi-user.target
