+++ extension/ss_client/ss_client@.service
@@ -1,0 +1,13 @@
+[Unit]
+Description=Firewalla SS Client
+After=network.target
+
+[Service]
+User=pi
+Type=simple
+Restart=always
+WorkingDirectory=/home/pi/firewalla/
+ExecStart=/bin/bash /home/pi/firewalla/extension/ss_client/ss_client.sh "%i"
+
+[Install]
+WantedBy=multi-user.target
